Zadanie 4.3 punkt 2
