+++ src/main/java/pl/javastart/task/Circle.java
@@ -14,4 +14,9 @@ class Circle {
     public void setR(int r) {
         this.r = r;
     }
+
+    double calculateCircleArea(Circle circle) {
+        final double pi = Math.PI;
+        return pi * Math.pow(r, 2);
+    }
 }
