+++ src/main/java/pl/javastart/task/Main.java
@@ -7,10 +7,9 @@ public class Main {
         Rectangle rectangle = new Rectangle(5, 10);
         Circle circle = new Circle(6);
         Triangle triangle = new Triangle(3, 4, 5);
-        ShapeCalculator shapecalculator = new ShapeCalculator();
-        System.out.println(shapecalculator.calculateSquareArea(square));
-        System.out.println(shapecalculator.calculateCircleArea(circle));
-        System.out.println(shapecalculator.calculateTrianglePerimeter(triangle));
-        System.out.println(shapecalculator.calculateRectPerimeter(rectangle));
+        System.out.println(square.calculateSquareArea(square));
+        System.out.println(circle.calculateCircleArea(circle));
+        System.out.println(triangle.calculateTrianglePerimeter(triangle));
+        System.out.println(rectangle.calculateRectPerimeter(rectangle));
     }
 }
