+++ src/main/java/pl/javastart/task/Triangle.java
@@ -35,4 +35,8 @@ class Triangle {
     public void setC(int c) {
         this.c = c;
     }
+
+    double calculateTrianglePerimeter(Triangle triangle) {
+        return a + b + c;
+    }
 }
